+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.3.4] - 2025-07-31
+ - Deafult message flush buffer size and time interval are set to zero (immediate flush
+ - Configuration stage is in logger launching not in the definition of logger thread
 
 ## [0.3.3] - 2025-07-22
  - Replace expect with unwrap for crossbeam try_send and write_all operations to avoid overhead